+++ AmazonChimeSDK/AmazonChimeSDK/internal/video/DefaultVideoClient.swift
@@ -24,7 +24,7 @@ class DefaultVideoClient: VideoClient {
         case Constants.errorLevel, Constants.fatalLevel:
             logger.error(msg: msg)
         default:
-            break
+            logger.default(msg: msg)
         }
     }
 }